+++ cmd/containerd/builtins/builtins.go
@@ -24,6 +24,7 @@ import (
 	_ "github.com/containerd/containerd/leases/plugin"
 	_ "github.com/containerd/containerd/metadata/plugin"
 	_ "github.com/containerd/containerd/pkg/nri/plugin"
+	_ "github.com/containerd/containerd/plugins/sandbox"
 	_ "github.com/containerd/containerd/plugins/streaming"
 	_ "github.com/containerd/containerd/plugins/transfer"
 	_ "github.com/containerd/containerd/runtime/restart/monitor"
