+++ services/sandbox/store_service.go
@@ -18,14 +18,15 @@ package sandbox
 
 import (
 	"context"
-	"errors"
 
 	"google.golang.org/grpc"
 
 	api "github.com/containerd/containerd/api/services/sandbox/v1"
+	"github.com/containerd/containerd/api/types"
+	"github.com/containerd/containerd/errdefs"
 	"github.com/containerd/containerd/log"
 	"github.com/containerd/containerd/plugin"
-	"github.com/containerd/containerd/services"
+	"github.com/containerd/containerd/sandbox"
 )
 
 func init() {
@@ -33,28 +34,21 @@ func init() {
 		Type: plugin.GRPCPlugin,
 		ID:   "sandboxes",
 		Requires: []plugin.Type{
-			plugin.ServicePlugin,
+			plugin.SandboxStorePlugin,
 		},
 		InitFn: func(ic *plugin.InitContext) (interface{}, error) {
-			plugins, err := ic.GetByType(plugin.ServicePlugin)
+			sp, err := ic.GetByID(plugin.SandboxStorePlugin, "local")
 			if err != nil {
 				return nil, err
 			}
-			p, ok := plugins[services.SandboxStoreService]
-			if !ok {
-				return nil, errors.New("sandbox store service not found")
-			}
-			i, err := p.Instance()
-			if err != nil {
-				return nil, err
-			}
-			return &sandboxService{local: i.(api.StoreClient)}, nil
+
+			return &sandboxService{store: sp.(sandbox.Store)}, nil
 		},
 	})
 }
 
 type sandboxService struct {
-	local api.StoreClient
+	store sandbox.Store
 	api.UnimplementedStoreServer
 }
 
@@ -67,25 +61,57 @@ func (s *sandboxService) Register(server *grpc.Server) error {
 
 func (s *sandboxService) Create(ctx context.Context, req *api.StoreCreateRequest) (*api.StoreCreateResponse, error) {
 	log.G(ctx).WithField("req", req).Debug("create sandbox")
-	return s.local.Create(ctx, req)
+	sb, err := s.store.Create(ctx, sandbox.FromProto(req.Sandbox))
+	if err != nil {
+		return nil, errdefs.ToGRPC(err)
+	}
+
+	return &api.StoreCreateResponse{Sandbox: sandbox.ToProto(&sb)}, nil
 }
 
 func (s *sandboxService) Update(ctx context.Context, req *api.StoreUpdateRequest) (*api.StoreUpdateResponse, error) {
 	log.G(ctx).WithField("req", req).Debug("update sandbox")
-	return s.local.Update(ctx, req)
+
+	sb, err := s.store.Update(ctx, sandbox.FromProto(req.Sandbox), req.Fields...)
+	if err != nil {
+		return nil, errdefs.ToGRPC(err)
+	}
+
+	return &api.StoreUpdateResponse{Sandbox: sandbox.ToProto(&sb)}, nil
 }
 
 func (s *sandboxService) List(ctx context.Context, req *api.StoreListRequest) (*api.StoreListResponse, error) {
 	log.G(ctx).WithField("req", req).Debug("list sandboxes")
-	return s.local.List(ctx, req)
+
+	resp, err := s.store.List(ctx, req.Filters...)
+	if err != nil {
+		return nil, errdefs.ToGRPC(err)
+	}
+
+	list := make([]*types.Sandbox, len(resp))
+	for i := range resp {
+		list[i] = sandbox.ToProto(&resp[i])
+	}
+
+	return &api.StoreListResponse{List: list}, nil
 }
 
 func (s *sandboxService) Get(ctx context.Context, req *api.StoreGetRequest) (*api.StoreGetResponse, error) {
 	log.G(ctx).WithField("req", req).Debug("get sandbox")
-	return s.local.Get(ctx, req)
+	resp, err := s.store.Get(ctx, req.SandboxID)
+	if err != nil {
+		return nil, errdefs.ToGRPC(err)
+	}
+
+	desc := sandbox.ToProto(&resp)
+	return &api.StoreGetResponse{Sandbox: desc}, nil
 }
 
 func (s *sandboxService) Delete(ctx context.Context, req *api.StoreDeleteRequest) (*api.StoreDeleteResponse, error) {
 	log.G(ctx).WithField("req", req).Debug("delete sandbox")
-	return s.local.Delete(ctx, req)
+	if err := s.store.Delete(ctx, req.SandboxID); err != nil {
+		return nil, errdefs.ToGRPC(err)
+	}
+
+	return &api.StoreDeleteResponse{}, nil
 }
